requirements.txt
# Core MCP dependencies
mcp>=1.0.0

# HTTP client
httpx>=0.24.0

# Configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Logging
structlog>=23.0.0